Pythagoras tree: Extracted tree logic to QGraphicsWidget
